+++ src/extension.ts
@@ -62,7 +62,7 @@ export function activate(context: ExtensionContext) {
     { command: "logtalk.make.reload",             callback: async (uri)  => LogtalkTerminal.makeReload(uri, linter)},
     { command: "logtalk.make.check",              callback: async (uri)  => LogtalkTerminal.makeCheck(uri, linter)},
     { command: "logtalk.run.tests",               callback: uri  => LogtalkTerminal.runTests(uri, linter)},
-    { command: "logtalk.run.doclet",              callback: uri  => LogtalkTerminal.runDoclet(uri)},
+    { command: "logtalk.run.doclet",              callback: uri  => LogtalkTerminal.runDoclet(uri, linter)},
     { command: "logtalk.scan.deadCode",           callback: uri  => LogtalkTerminal.scanForDeadCode(uri, deadCodeScanner)},
     { command: "logtalk.generate.documentation",  callback: uri  => LogtalkTerminal.genDocumentation(uri, documentationLinter)},
     { command: "logtalk.generate.diagrams",       callback: uri  => LogtalkTerminal.genDiagrams(uri)},